add jwt token verify as middle ware
